Altered the apple coordinate update method.
When updating the apple coordinate, I now give the apple a list of positions which are occupied by the snake as a parameter. This is generated by a new method added to the snake. This way the apple only sees what it needs from the snake and not the whole snake.